actualizacion

+++ Haskel/ejercicio1.hs
@@ -3,5 +3,5 @@ sumaDos x = x + 2
 msDos x y = sumaDos x + sumaDos y
 condicion x = if x > 100 then x else x*2
 
-sumaTres :: Int -> Int -> Int -> Int
+sumaTres :: Float -> Float -> Float -> Float
 sumaTres x y z = x + y + z
